refactor(trainer): WIP refactor
