+++ optimum/neuron/__init__.py
@@ -27,8 +27,7 @@ from transformers.utils import _LazyModule
 
 _import_structure = {
     "hf_argparser": ["NeuronHfArgumentParser"],
-    "trainers": ["NeuronTrainer", "Seq2SeqNeuronTrainer", "NeuronSFTTrainer", "NeuronORPOTrainer"],
-    "training_args": ["NeuronTrainingArguments", "Seq2SeqNeuronTrainingArguments"],
+    "trainers": ["NeuronTrainer", "Seq2SeqNeuronTrainer", "NeuronSFTTrainer", "NeuronORPOTrainer", "NeuronTrainingArguments", "Seq2SeqNeuronTrainingArguments", "NeuronSFTConfig", "NeuronORPOConfig"],
     "modeling_traced": ["NeuronTracedModel"],
     "modeling": [
         "NeuronModelForFeatureExtraction",
@@ -73,7 +72,6 @@ _import_structure = {
         "NeuronPartialState",
     ],
     "pipelines": ["pipeline"],
-    "utils": ["NeuronSFTConfig", "NeuronORPOConfig"],
 }
 
 # Model structures
@@ -141,9 +139,7 @@ if TYPE_CHECKING:
     from .models.inference.whisper import NeuronWhisperForConditionalGeneration
     from .models.inference.yolos import NeuronYolosForObjectDetection
     from .pipelines import pipeline
-    from .trainers import NeuronORPOTrainer, NeuronSFTTrainer, NeuronTrainer, Seq2SeqNeuronTrainer
-    from .training_args import NeuronTrainingArguments, Seq2SeqNeuronTrainingArguments
-    from .utils import NeuronORPOConfig, NeuronSFTConfig
+    from .trainers import NeuronORPOTrainer, NeuronSFTTrainer, NeuronTrainer, Seq2SeqNeuronTrainer, NeuronTrainingArguments, Seq2SeqNeuronTrainingArguments, NeuronORPOConfig, NeuronSFTConfig
 
 else:
     import sys
