+++ optimum/neuron/pipelines/transformers/__init__.py
@@ -13,6 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .base import (
-    pipeline,
-)
+from .transformers import
