+++ optimum/neuron/trainers/__init__.py
@@ -1,0 +1,20 @@
+# coding=utf-8
+# Copyright 2025 The HuggingFace Inc. team. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .transformers import NeuronTrainer, Seq2SeqNeuronTrainer
+from .sft_trainer import NeuronSFTTrainer
+from .orpo_trainer import NeuronORPOTrainer
+from .training_args import NeuronTrainingArguments, Seq2SeqNeuronTrainingArguments
+from .trl_utils import NeuronSFTConfig, NeuronORPOConfig
